# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ghana Pass API",
    "description": "National Digital Identity Platform — API Documentation\n\nGhana Pass provides a centralized identity verification and authentication infrastructure for Ghana's digital ecosystem.",
    "version": "1.0.0",
    "contact": { "name": "Ghana Pass Support", "email": "support@ghanapass.gov.gh" }
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Development" },
    { "url": "https://api.ghanapass.gov.gh", "description": "Production" }
  ],
  "tags": [
    { "name": "Authentication", "description": "User registration and login" },
    { "name": "Identity Verification", "description": "Ghana Card and facial verification" },
    { "name": "SIM Verification", "description": "SIM card ownership verification" },
    { "name": "Documents", "description": "Digital document signing" },
    { "name": "OAuth", "description": "OAuth 2.0 / OpenID Connect" },
    { "name": "User", "description": "User profile and activity" },
    { "name": "Admin", "description": "Administrative operations" },
    { "name": "Developer", "description": "Developer portal and app management" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ghana_card_number", "phone", "full_name", "date_of_birth", "password"],
                "properties": {
                  "ghana_card_number": { "type": "string", "example": "GHA-123456789-1" },
                  "phone": { "type": "string", "example": "+233240000001" },
                  "full_name": { "type": "string", "example": "Kwame Asante" },
                  "date_of_birth": { "type": "string", "example": "1990-01-15" },
                  "password": { "type": "string", "example": "securepass123" },
                  "email": { "type": "string", "example": "kwame@email.com" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration successful" },
          "409": { "description": "Ghana Card already registered" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login with Ghana Card",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ghana_card_number", "password"],
                "properties": {
                  "ghana_card_number": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT tokens" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/verify/identity": {
      "post": {
        "tags": ["Identity Verification"],
        "summary": "Verify identity via Ghana Card",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ghana_card_number": { "type": "string", "example": "GHA-123456789-1" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Verification result with NIA data" }
        }
      }
    },
    "/api/verify/face": {
      "post": {
        "tags": ["Identity Verification"],
        "summary": "Facial recognition verification",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": { "type": "string" },
                  "selfie_data": { "type": "string", "description": "Base64 encoded selfie" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Face verification result with liveness and match scores" }
        }
      }
    },
    "/api/sim/verify": {
      "post": {
        "tags": ["SIM Verification"],
        "summary": "Verify SIM card ownership",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone_number", "ghana_card_number"],
                "properties": {
                  "phone_number": { "type": "string", "example": "+233240000001" },
                  "ghana_card_number": { "type": "string", "example": "GHA-123456789-1" },
                  "provider": { "type": "string", "example": "MTN" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "SIM verification result" }
        }
      }
    },
    "/api/documents/sign": {
      "post": {
        "tags": ["Documents"],
        "summary": "Digitally sign a document",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["document_id", "pin"],
                "properties": {
                  "document_id": { "type": "string" },
                  "pin": { "type": "string", "example": "1234" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Signature details with hash" }
        }
      }
    },
    "/api/oauth/authorize": {
      "get": {
        "tags": ["OAuth"],
        "summary": "OAuth authorization endpoint",
        "parameters": [
          { "name": "client_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "redirect_uri", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "response_type", "in": "query", "required": true, "schema": { "type": "string", "enum": ["code"] } },
          { "name": "scope", "in": "query", "schema": { "type": "string" } },
          { "name": "state", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Client info for consent screen" }
        }
      }
    },
    "/api/oauth/token": {
      "post": {
        "tags": ["OAuth"],
        "summary": "Exchange authorization code for tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grant_type": { "type": "string", "enum": ["authorization_code", "refresh_token"] },
                  "code": { "type": "string" },
                  "client_id": { "type": "string" },
                  "client_secret": { "type": "string" },
                  "redirect_uri": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Access token, refresh token, and ID token" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}